+++ src/main/java/util/events/testevents/UILogicRefreshEvent.java
@@ -1,4 +1,6 @@
-package util.events;
+package util.events.testevents;
+
+import util.events.Event;
 
 public class UILogicRefreshEvent extends Event {
     public UILogicRefreshEvent() {}
