+++ src/main/java/util/events/testevents/UILogicRefreshEventHandler.java
@@ -1,6 +1,7 @@
-package util.events;
+package util.events.testevents;
 
 import com.google.common.eventbus.Subscribe;
+import util.events.EventHandler;
 
 @FunctionalInterface
 public interface UILogicRefreshEventHandler extends EventHandler {
