+++ src/main/java/util/events/testevents/UpdateDummyRepoEvent.java
@@ -1,4 +1,6 @@
-package util.events;
+package util.events.testevents;
+
+import util.events.Event;
 
 public class UpdateDummyRepoEvent extends Event {
